+++ src/neuroconv/datainterfaces/__init__.py
@@ -34,6 +34,7 @@ from .ecephys.kilosort.kilosortdatainterface import KiloSortSortingInterface
 from .ecephys.edf.edfdatainterface import EDFRecordingInterface
 from .ecephys.tdt.tdtdatainterface import TdtRecordingInterface
 from .ecephys.plexon.plexondatainterface import PlexonRecordingInterface
+from .ecephys.biocam.biocamdatainterface import BiocamRecordingInterface
 
 # Icephys
 from .icephys.abf.abfdatainterface import AbfInterface
@@ -86,6 +87,7 @@ interface_list = [
     EDFRecordingInterface,
     TdtRecordingInterface,
     PlexonRecordingInterface,
+    BiocamRecordingInterface,
     # Icephys
     AbfInterface,
     # Ophys